+++ src/main/java/com/cryptoArb/domainOld/PriceTick_old.java
@@ -1,15 +1,17 @@
 package com.cryptoArb.domainOld;
 
+import java.math.BigDecimal;
+
 public class PriceTick_old {
 
 
     private final CurrencyPair_old pair;
     private final Exchange_old exchangeOld;
     private final long timestamp;
-    private final double bidPrice;
-    private final double askPrice;
+    private final BigDecimal bidPrice;
+    private final BigDecimal askPrice;
 
-    public PriceTick_old(CurrencyPair_old pair, Exchange_old exchangeOld, long timestamp, double bidPrice, double askPrice) {
+    public PriceTick_old(CurrencyPair_old pair, Exchange_old exchangeOld, long timestamp, BigDecimal bidPrice, BigDecimal askPrice) {
         this.pair = pair;
         this.exchangeOld = exchangeOld;
         this.timestamp = timestamp;
@@ -29,11 +31,11 @@ public class PriceTick_old {
         return timestamp;
     }
 
-    public double getBidPrice() {
+    public BigDecimal getBidPrice() {
         return bidPrice;
     }
 
-    public double getAskPrice() {
+    public BigDecimal getAskPrice() {
         return askPrice;
     }
